Commit 5 : __________
modified:   class/Runner.cpp
modified:   hdr/MyRunner_v2.hpp
modified:   hdr/Runner.hpp

Aliases are functionnal. They must be defined in the following file :
			"$HOME/.config/MyRunner_v2/aliases".

Lines, which begin with a sharp '#' are commented.
An alias is defined so :
			alias="command with its arguments"

For the moment, the "aliases" file must be created and the aliases must be
wrotten through a text editor in the configuration file.
In the next commit, the options 'add alias', 'remove alias' and
'show the "aliases" file' will be added to the "MyRunner_v2".